requirements.txt
yara-python
psutil
scapy
faker
requests
paramiko
pytest
